last commit

+++ src/main/java/org/pago/ggdrop/entity/Item.java
@@ -18,7 +18,7 @@ public class Item {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     @Column(nullable = false, length = 30)
     private String name;
@@ -42,7 +42,7 @@ public class Item {
         this.price = price;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
